Move Source class from buildout.py to base.py.
I want to use this in pip.py as well.
